fix: eopsode pollution
